Cambié el tipo de dato de reputación de float a double

+++ public_html/Migrations/profileTable.php
@@ -14,7 +14,7 @@ if ($conn->connect_error) {
 $sql = "CREATE TABLE profile (
 id INT(6) UNSIGNED AUTO_INCREMENT PRIMARY KEY,
 name VARCHAR(150) NOT NULL,
-reputation FLOAT(50) NOT NULL,
+reputation DOUBLE(50) NOT NULL,
 status BOOLEAN NOT NULL,
 reg_date TIMESTAMP
 )";
